#232 Adds generic object pooling based on reference counting.
This implementation of object pooling is based on PR #257. It is thread-safe, uses a BitSet for performance and an internal List. In the Netty implementation of Recycler, the object pools are ThreadLocal and an array and counting is managed via ++ and --. This implementation differs from the Netty version in that the object pool can be shared across threads and reference counting is done. When the reference count for an allocated object reaches 0 it is recycled.