#232 Integrated Allocator into the Dispatchers.
This is a very pervasive changeset as it touches nearly every part of the internal workings of all the Dispatchers. Since the inclusion of a more powerful Allocator abstraction in recent commits, it seemed to make sense to re-write the Dispatchers to make use of this to provide some consistent efficiencies. There's still some work to do and bugs to fix so this shouldn't be considered a complete commit. But it's necessary for others working on this branch to have the latest changes.