#287 Refactored TCP support to use common abstractions called "NetXXX" and added UDP support.
UDP support was not difficult to implement but we needed to use the same abstractions as the TCP support. This involved renaming and refactoring most classes within the reactor-tcp submodule, including the module itself, now called 'reactor-net'.

Functionality that's common to both client and server are abstracted out into AbstractNetPeer and AbstractNetChannel. Specific connection details are left up to the implementations of TcpClient/TcpServer/DatagramServer.