Iterator Publisher
Previously, only an Iterable and not an Iterator could be used as a
publisher.  As the Iterable was used to call .iterator() there should
have been an equivalent for passing in the Iterator directly.  This
change abstracts a common class for the actual iteration while
creating separate child classes for initializing the context from an
Iterable and and Iterator.  It also adds from() and convert() behavior
to Publishers as well as from() behavior to Streams.