Start moving Bus in the direction of Firehose.
Make Bus interface more generic, allow alternative
implementations of Bus that do not have event.

Pull up the second overload for `notify` to Bus
interface.

Extract AbstractBus as a base to something that 
used to be Firehose, and will be Bus without event
now.

Registry now understands the "Firehose" way of 
subscribing to the objects which doesn't break
interface.

Make it possible to consume both keys and values
(feature required for Pipes)