Avoid implicit casting in ProcessorGroup
ProcessorGroup#createBarrier creates ProcessorBarriers using the generic
type T. This type is not always desired, e.g. when creating dispatchers
for other types. Casts can easily be avoided and thus problems with the
Eclipse compiler by adding a new generic type to
ProcessorGroup#createBarrier.

This is a follow up to PR 587 [1].

[1] https://github.com/reactor/reactor/pull/587