zip and zipWith to Tuples
Previously, only the zip(Publisher, Publisher) method would create a
Stream of Tuples.  This change updates all of the zip(...) methods
with a variant that creates a Stream of the Tuples already used
internally.  In addition, variants of the zipWith(...) methods that
create Streams of Tuples were created as well.