[backport] Rationalize version properties, add groupId task
The version properties don't need to be repeated for each artifact of
a given repo. Furthermore, if they reflect the name of the repos more
directly it will benefit the releaser script we a considering using.

The groupId task was also added in preparation for the releaser script.

(cherry picked from commit 5386fee659b93ff5e44841314d408eb5fe7e0d86)