[build] master is 2020.0 Europium, introduce new versioning scheme
This commit switches the effort on master from Dysprosium to Europium.

It also introduces a new versioning scheme, using -SNAPSHOT
instead of .BUILD-SNAPSHOT as the qualifier for snapshots.
The whole scheme is detailed in the reference documentation of core,
as well as in the README here.

Adaptations are made to the build script to accommodate it.